Upgrade to latest Scala/SBT and add tests for Thrift `0.12.0`.
This commit adds tests using the SBT scripting framework
(https://www.scala-sbt.org/1.x/docs/Testing-sbt-plugins.html) that
ensure that Thrift sources get compiled correctly with the plugin.
Curently it only tests Java but that's okay for testing at least the
basic framework for now.

I have bumped all of the versions of everything at the same time - Scala
to `0.12.8`, SBT to `1.2.8`. In addition, we'll now test on CI with
Thrift `0.12.0` which is the latest release.